test: 상품 도착 확인 이벤트 컨슈머 통합 테스트 구현
