+++ src/test/java/com/trustamarket/inspectionservice/inspection/adapter/in/messaging/CenterSlotAssignedConsumerTest.java
@@ -1,63 +1,15 @@
 package com.trustamarket.inspectionservice.inspection.adapter.in.messaging;
 
 import com.trustamarket.inspectionservice.inspection.application.dto.command.RequestInspectionCommand;
-import com.trustamarket.inspectionservice.inspection.application.port.in.RequestInspectionUseCase;
-import org.junit.jupiter.api.BeforeEach;
+import com.trustamarket.inspectionservice.support.KafkaConsumerIntegrationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.kafka.config.KafkaListenerEndpointRegistry;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.test.EmbeddedKafkaBroker;
-import org.springframework.kafka.test.context.EmbeddedKafka;
-import org.springframework.kafka.test.utils.ContainerTestUtils;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.timeout;
 
-@SpringBootTest
-@EmbeddedKafka(
-        partitions = 1,
-        topics = {"product.inspection-requested", "center.slot-assigned"},
-        bootstrapServersProperty = "spring.kafka.bootstrap-servers"
-)
-@TestPropertySource(properties = {
-        "spring.cloud.config.enabled=false",
-        "eureka.client.enabled=false",
-        "spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1",
-        "spring.datasource.driver-class-name=org.h2.Driver",
-        "spring.datasource.username=sa",
-        "spring.datasource.password=",
-        "spring.jpa.database-platform=org.hibernate.dialect.H2Dialect",
-        "spring.jpa.hibernate.ddl-auto=create-drop",
-        "spring.jpa.properties.hibernate.default_schema="
-})
-@DirtiesContext
-class CenterSlotAssignedConsumerTest {
-
-    @Autowired
-    private KafkaTemplate<String, String> kafkaTemplate;
-
-    @Autowired
-    private KafkaListenerEndpointRegistry registry;
-
-    @Autowired
-    private EmbeddedKafkaBroker embeddedKafkaBroker;
-
-    @MockitoBean
-    private RequestInspectionUseCase requestInspectionUseCase;
-
-    @BeforeEach
-    void waitForConsumerAssignment() throws Exception {
-        for (var container : registry.getListenerContainers()) {
-            ContainerTestUtils.waitForAssignment(container, embeddedKafkaBroker.getPartitionsPerTopic());
-        }
-    }
+class CenterSlotAssignedConsumerTest extends KafkaConsumerIntegrationTest {
 
     private static final String VALID_PAYLOAD = """
             {
